dev-requirements.txt
django-pdb
django-extensions
django-debug-toolbar

-r requirements.txt